+++ BiliAPI/BiliDynamic/DynamicModel/BiliDynamicUserInfo.cs
@@ -4,7 +4,7 @@ namespace BiliAPI.BiliDynamic.DynamicModel
 {
     public struct BiliDynamicUserInfo
     {
-        public int? uid { get; set; }
+        public long? uid { get; set; }
         public string? uname { get; set; }
         public string? face { get; set; }
         public int? face_nft { get; set; }
